+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [feature] `shelly config create` will check if config already exists in
   specified path
+* [bugfix] With multiple clouds in `Cloudfile`, `shelly config create`
+  should not open the editor if no cloud is specified
 
 ## 0.4.29 / 2014-03-27